Tweaked Fix for Reloading when Clicking Students Nav Button Again
I changed the check for whether or not to reset the page slightly. It no longer incorrectly resets the page when loading two profiles one after the other. However, it does incorrectly reset the page when loading the same profile twice.

The root of this issue is that the Router, which controls the Nav Button, doesn't update any reliable state or props information, so it's very hard to differentiate between a re-render caused by something in Students.js or one caused by clicking the Students Nav Button.