Fixed an issue with asynchronous POSTs and GETs
It was possible that a newly edited student wouldn't finish POST-ing its new fields before the GET was sent for their information. This didn't cause any issues displaying data, since we kept a local copy. However, it meant that we didn't have the ids that the database generated, so if we went to immediately delete that student we missed some rows.